ipsec-wrapper: Pass secure-upstreams explicitly
Adding 'secure-upstreams' to the service interface for the ipsec-wrapper
is exactly the kind of "merge together definitions from an attribute in
a global namespace" thing that underlies one of the major difficulties
with the nixos module system. Any functions that add in some extra
default services which themselves have configuration options should
expose those through top-level function arguments instead.

Note that ipsec-wrapper.nix is approximately such a function, though due
to its nature as a bridge between the module system it exposes its
arguments modularly (which is to say, non-modularly).