Introduce run-as-user for user switching
By default we compile a wrapper that calls setuid before execing the
program, but the program can specialize running as another user.
multiplex-activations specializes it to run the *wrapped* program as the
new user (so that resources such as sockets are acquired as root) and
execve specializes it to call setuid in the same process.